[PR-236] duplicate bank accounts data drawer (#12605)
## Describe your changes

Essentially a direct duplicate of the `card-duplicate-drawer` - almost all of the functionality is shared but we generally prefer to be more verbose and simpler / more readable.

![Screenshot 2024-10-08 at 12.54.20 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/1975f962-0dfc-4d11-bc03-3170f8187ac5.png)

Basically all of the same filtering logic, etc. The only real diffs here are (a) the text in the drawer and (b) using bank account DIs in the test config.

## Caveats

This still does not show to users!

## Test Plan

In UI and `bank-duplicate-drawer.test.tsx`