Add property-based tests for VDR (#12578)
## Describe your changes
Compares `footprint-dr`'s view of the data against ground truth models
of vaults as KV maps. I validated that this test found some of the
recent bugs I've fixed, but it doesn't test failure paths well. To do
that, we could inject failures via additional VDR worker knobs. However,
along with the formal modeling of the design, this first pass of
randomized testing gives me enough confidence in the implementation that
I think we can call it done for now.

## Caveats


## Test Plan