Support making adhoc document requests for users without a WF (#12712)
## Describe your changes
Request from wingspan:
https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1728578633966439?thread_ts=1728506545.933679&cid=C05U1CAD6FQ

Very silly, we have this technical constraint that a workflow _must_ be
associated with a playbook. Bifrost has a few codepaths where it fetches
the playbook in order to display general, non-onboarding-specific
information like "Acme bank has requested you to upload a document" or
something. So we would never allow requesting a document from users who
didn't already onboard onto a playbook since we would then have no
playbook to associate with the workflow.

This relaxes the constraint with a hack... we just choose a playbook
owned by the tenant to randomly attach to the WorkflowRequest (and thus
the Workflow).

Not the best work... but it does the trick.

## Caveats


## Test Plan