[PR-277] useFilters handling for is_sentilink param (#12811)
## Describe your changes

Title! We currently control whether or not to open the drawer based on query params. We will be opening a different drawer in the case that this is a sentilink signal, so passing it as q uery parameter. Want to not add too much so we'll only include if it is s a sentilink signal (i.e. - if the `is_sentilink` parameter isn't included, we will assume it is false).

In UI:

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/XsTtzwNbP0GdQMHxY1hw/0091cf37-a629-4c01-a7f9-f4adc3eb9262.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/XsTtzwNbP0GdQMHxY1hw/0091cf37-a629-4c01-a7f9-f4adc3eb9262.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/0091cf37-a629-4c01-a7f9-f4adc3eb9262.mov">Screen Recording 2024-10-15 at 5.16.19 PM.mov</video>

## Caveats

## Test Plan

I unskipped all of the risk-signals-list tests, except for three that remain broken, and added three new tests to make sure this logic continues to work