[PR-278] sentilink-details scaffolding (#12821)
## Describe your changes

Title! Generally just replicate the shape of standard-details and get general design / structure here. We don't show this anywhere but it will let me dev a bunch of the individual components in the meantime. Also include the base for `Content` here, `Loading` and `Error` will come in the future.

Note that because of this fetch, we need to mock the requests to the sentilink endpoint in risk-signals-details now.

Underlying fetching logic is now called here and means we can use it. If I override the `isOpen` in both `StandardDetails` and `SentilinkDetails`, I can test other things here without worrying about users seeing it.
<img width="1336" alt="Screenshot 2024-10-16 at 9 06 17 AM" src="https://github.com/user-attachments/assets/cf6438c9-bc6a-46b2-a0af-82841feaf590">

## Caveats

N/A

## Test Plan

N/A - doesn't show to users yet
In dev, see screenshot above