[PR-272] risk indicator with tests  (#12825)
## Describe your changes

More and less fraudy indicator. This is purely functional, because we have to change the icons anyways I'm planning to do a separate PR to double check all of the styling (although it should be good with the mocks now).

Here's what it looks like with two of them in UI, icons will change
<img width="463" alt="Screenshot 2024-10-16 at 10 47 45 AM" src="https://github.com/user-attachments/assets/3074cea4-26e2-4d9d-88d7-3e89ce164524">

## Caveats
Doesn't show up anywhere / will be used in one of these components, we need to swap the icons in the future.

## Test Plan

`risk-indicator.test.tsx`