[PR-274] score component (#12836)
## Describe your changes

Actual score UI now! Claudio says looks good:

![Screenshot 2024-10-16 at 12.03.39 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/4649266e-aa78-4068-a338-8e50b741135a.png)

## Caveats

There's a BE change coming for what color we use to display the score, plus we'll also use the pending color. This still hasn't merged so it will come in another PR. Also, 'score' will be swapped for 'Synthetic score' and 'ID theft score' / passed as a prop, in the next PR.

## Test Plan

`score.test.tsx`