[PR-292] redesign with moreFraudy / lessFraudy section headers (#12843)
## Describe your changes

Here's the UI PR! Because we changed the designs midway through eng (after realizing that we wanted to display the full codes / they were getting cutoff before), we have this. Big changes:
- No risk indicator in `reason-codes` / it's now a component of scores
- Grouped sections with all reason codes of a certain score in scores (+ tests)

![Screenshot 2024-10-16 at 2.07.51 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/cf702e40-c439-41d9-a901-7bfb8fc1e462.png)

## Caveats

There's a PR that will come after this with a few more style changes and the correct down icon.

## Test Plan

`scores.test.tsx`