[PR-279] loading shimmer component (#12846)
## Describe your changes

Title! Pretty self-explanatory. Here's a video of how this looks when we throttle connection:

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/XsTtzwNbP0GdQMHxY1hw/31f2d894-9b1e-4af0-9124-f17a8063a4a7.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/XsTtzwNbP0GdQMHxY1hw/31f2d894-9b1e-4af0-9124-f17a8063a4a7.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/31f2d894-9b1e-4af0-9124-f17a8063a4a7.mov">Screen Recording 2024-10-16 at 2.41.29 PM.mov</video>

## Caveats

N/A

## Test Plan

In UI. It's very difficult to elegantly test the `sentilink-details` component right now because we ensure that it is never open (there's literally an `&& false` in the component for it to be open). It's possible unit tests can come for this in a future PR, although they'd have to fake the `isPending` from react-query, and I'm not sure how difficult that is / if it is worth it. Generally I don't think this is so high risk and it works great in UI so I think we are probably fine in the case it's a lot of lift / not worth it.