[PR-269] StreetView loading spinner (#12793)
## Describe your changes

Add loading state to Street View. Unfortunately, because some of the
load is handled by Google, we (Rodrigo and I, after spending about an
hour on it) couldn't figure out a good/reliable way to keep our spinner
up while the black screen is there. However, this does pretty well for
the middle! Pedro is OOO but this is based on his mocks, also seems like
it isn't too big of a deal.



https://github.com/user-attachments/assets/a16262df-0815-4e2d-bd4a-644c5855e86e


<img width="524" alt="Screenshot 2024-10-17 at 7 39 34 AM"
src="https://github.com/user-attachments/assets/4ba64601-b7ef-4222-a424-017768d1f2ef">

## Caveats

N/A

## Test Plan

In UI

---------

Co-authored-by: rodrigopagnuzzi <rodrigo@onefootprint.com>