Add script for end-to-end VDR validation (#12879)
## Describe your changes
This is approximately the same way that Grid will validate VDR. I
implemented this as an additional test that all the VDR correctness work
has been successful end-to-end.

## Caveats
The script isn't perfect due to the `can_access_data` OBC setting
preventing decryption of some fields. I spent some time evaluating how
much work it is to disable that feature, and it's a decent amount (a few
days probably). For now, I just ignored decryption mismatches that are
consistent with the current OBC `can_access_data` behavior.

## Test Plan
I ran this on the dev sandbox integration test tenant, and didn't find
any issues.