[PR-297] dont do the sentilink fetch if in standard details (#12908)
## Describe your changes

Title! Currently we are getting some errors because the sentilink
fetches are calling even when we are looking at non-sentilink signals.
This is because we aren't making the call to the hook conditional. Now
it is, so we don't have this problem:

Before:

![Screenshot 2024-10-17 at
8.55.59 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/28c85524-e06c-4785-aa61-f58a6c0e927e.png)

After

![Screenshot 2024-10-17 at
8.57.37 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/9d80df0b-88ae-4cdb-943e-bcf473b3df58.png)

## Caveats

We need to disable the standard fetch when we are doing sentilink
details, but right now `StandardDetails` covers all cases so we'd be
breaking the sentilink case if we did that now. I've ticketed this in
[PR-296](https://linear.app/footprint/issue/PR-296/dont-do-the-standard-fetch-if-in-sentilink-details)
and will add.


## Test Plan

In UI

---------

Co-authored-by: Rafael Michels Motta <rafael@onefootprint.com>