Update VDR to avoid stalling on retroactive DL writes (#12917)
## Describe your changes
Currently, if we retroactively create DLs at existing scoped vault
version that is already backed up, new scoped vault versions for the
scoped vault will never select those DLs for backup. This means the new
SVVs will never be selected as "complete," so the VDR backup will not
make progress on that scoped vault. If this happens to enough vaults (>=
the manifest batch size), VDR as a whole will not make progress for that
tenant/is_live.

To fix, we can relax the DL selection query to select retroactively
created DLs as well. This also makes it simpler to do backfills if we
don't need VDR to update already-written records. In that common case,
we don't need to null out the `backed_up_by_vdr_config_id` column or
stop the VDR worker.

## Caveats


## Test Plan
existing property based tests and updated model checked:
```
Model checking completed. No error has been found.
  Estimates of the probability that TLC did not check all reachable states
  because two distinct states had the same fingerprint:
  calculated (optimistic):  val = .0012
  based on the actual fingerprints:  val = .0042
307276866 states generated, 112009400 distinct states found, 0 states left on queue.
The depth of the complete state graph search is 60.
The average outdegree of the complete state graph is 1 (minimum is 0, the maximum 12 and the 95th percentile is 3).
Finished in 01h 53min at (2024-10-17 14:49:01)
```