[1/4][BE-1708] Add new workflow_request_junction table and start writing (#12928)
## Describe your changes
We're going to start having more use cases where a `WorkflowRequest` is
tied to a business. For the only existing case, it was sufficient to
just save an `fp_bid` inside the WFR config. Now that we'll also have
"redo KYB" flows, we need a more general way of linking a WFR to a
scoped business vault.

This makes a new junction table to tie WFR <> SV, and the idea is that
we could have two rows in the junction table for each WFR: one to tie it
to a user, and one to tie it to a business.

This will also help us to better support filtering on "info requested"
from a business too

## Caveats


## Test Plan