[PR-288] score color band support (#12876)
## Describe your changes

Dave wants to control the color we show to users for the score from the BE, this PR gets rid of the logic we had on the frontend, deals with all of the typing for this new `score_band` field, and creates logic +tests in `score` to make sure it all works.

![Screenshot 2024-10-16 at 3.28.11 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/e1222e54-5f21-4a2c-860a-d60c47274079.png)

## Caveats

N/A

## Test Plan

`score.test.tsx, manually changing these in UI, and CI/typecheck.