[PR-290] sentilink details logic shows properly + tests + put into prod (#12878)
## Describe your changes

Title! This is just the logic to put details in in a mutually exclusive way! Lots of testing as well, far fewer changes than it looks, large just because i'm moving a bunch of old tests into a describe block.

## Caveats

We need to do the redesign now

## Test Plan

`standard-details.test.tsx` and `sentilink-details.test.tsx`