[1/n][BE-1713] Write wfr_junction for businesses (#12963)
## Describe your changes
Since we're going to have multiple kinds of WFR configs that are
associated with a business, we need to store the business <> WFR
relationship outside of the config. We made the `wfr_junction` table in
the PRs below this to serve this purpose. Now, instead storing the
associated business inside the wfr config (specifically in the document
config), we're going to create a `wfr_junction` row to associate the WFR
with the business.

This also starts to accept an `fp_bid` in the body of the `POST
/entities/<>/actions` - I will update the dashboard in a future PR to
send the `fp_bid` in this new, more general location

## Caveats


## Test Plan