[PR-299] "No other BI" dialog for when ownership percentage doesn't add up to 100 (#12995)
## Describe your changes

Title! Bruno is working on the logic here already / we realized we
collided here, but the UI is done! Logic will come in a future PR.

"Add another" change tested 

![Screenshot 2024-10-18 at
11.04.19 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/13dff10e-86a9-4b3b-9af1-7e663cc53654.png)

![Screenshot 2024-10-18 at
10.58.26 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/8342c910-1462-474f-862c-d8ac9331f9c4.png)

![Screenshot 2024-10-18 at
10.58.10 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/2da1c111-c6b0-4734-ac00-cfed076931da.png)

## Caveats

See above. No logic in this PR. Also no spanish in the translations, not
sure how this works.

## Test Plan

N/A - this UI is used nowhere - this