[bifrost][BE-1715] Handle `recollectAttributes` in KYB (#12996)
## Describe your changes
In the KYC + KYB requirements, we have `missingAttributes` and
`populatedAttributes`. Normally, we skip showing the individual
collection screens for data in `populatedAttributes`, but we have some
cases where we want to explicitly show a data collection screen for a
field in `populatedAttributes`.

So, this adds a third field to these requirements:
`recollectAttributes`. `recollectAttributes` will be a subset of
`populatedAttributes`, and when they exist, bifrost should show the
individual collection screen for those attributes, even though the data
is already populated.

This implements `recollectAttributes` only for the KYB machine, since
the dashboard won't be creating workflows with `recollectAttributes` for
KYC workflow yet


## Caveats


## Test Plan
Unit test, tested locally