[FE-723] deprecate decryptedAttributes in frontend (#13032)
## Describe your changes

Title! This is mostly just removing the type and removing this stuff
from tests. It's also removed from the `Entity` type so people won't be
able to use it in the future.


I also tested on `row.tsx` that we change to make sure it works, which
it does:

![Screenshot 2024-10-21 at
9.50.02 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/b495e60f-4b0e-4552-9d75-a2e39160d2a8.png)

## Caveats

I've dug up some stuff we are doing in the mobile package that uses the
`attributes`, `decryptedAttributes`, `decryptableAttributes` API, there
will be a future PR that deals with those.

## Test Plan

Existing testing and typecheck.