[1/2] Add UUID column to beneficial_owner (#13087)
## Describe your changes
In order to make the frontend a little cleaner, we're going to support
allowing the frontend to generate a UUID for each new beneficial owner,
rather than relying on the backend-generated BoLinkId. This allows us to
support an `update_or_create` operation instead of separate `update` and
`create` operations

## Caveats
I could name this `idempotency_uuid` or `idempotency_id`? wdyt

## Test Plan