Improve perf of get_complete_svvs_for_svv_batch through a different join strategy (#13213)
## Describe your changes
Both the original query and the attempted fix in #13057 do not work
since their plans attempt to materialize the whole vault_dr_blob table
for the given config ID. Since the `vault_dr_blob` table is huge, this
makes the queries intractably slow.

I found through a ton of trial and error that it was likely the
`dl.created_at <= svv.seqno` part of the query that was correlated with
this materialize. By expanding the SVV batch to all versions of the
scoped vaults in the batch up to and including the versions in the
batch, and then joining on `dl.created_at = svvs_for_same_sv.seqno`, we
avoid the huge materialize. The approach has been semi-vetted in
production already by way of
https://github.com/onefootprint/monorepo/pull/13080, which does a
similar style of join.

While updating the TLA+ specs accordingly to double check the logic, I
improved the tooling a bit so we don't have to manually tweak the model
checking config as much to test different parameters.

## Caveats


## Test Plan