[PR-315] hidden security-logs page (#13223)
## Describe your changes

![Screenshot 2024-10-24 at 12.31.00 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/e6eb4dd6-4cd5-43c8-bef6-865e8a275e61.png)

Title! We have this `security-logs-private` page that I'll build everything on. It doesn't add a tab or anything / we'll return private for non-firm-employees, and can work / test here without disrupting anything. I imagine, when we merge everything later, we can just delete the `security-logs` folder and rename this one.

## Caveats

N/A

## Test Plan

In UI