[PR-316] factor security logs hooks into common hooks (#13225)
## Describe your changes

This is just a bunch of file moves so that we can use these in `security-logs-private`. I don't think these utils should live in the common dashboard hooks, because they are only scoped to security logs, but I think this is a better solution (to move them to common, and move them back after we deprecate the old `security-logs`) than copying all of the hooks. I've added a ticket to the end to move these back, [PR-320: move security logs hooks back into security-logs / away from dashboard common hooks](https://linear.app/footprint/issue/PR-320/move-security-logs-hooks-back-into-security-logs-away-from-dashboard).

## Caveats

No change to users - this is just: my adding an index file, and a bunch of file moves.

## Test Plan

CI