[PR-326] insight event hover logic (#13282)
## Describe your changes

Title! I anticipate this one being a bit more difficult, this is to deal with a hover card we have in the mocks, which look like this:

![Screenshot 2024-10-25 at 12.34.17 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/4ea10c0f-71c5-4157-a9d3-110b110d6a01.png)

## Caveats

I haven't built the actual `UserInsights` card yet, that's coming. There will also be some interesting alignment stuff when this is actually put on a timeline, which I've yet to deal with, want this to be mostly functional and to confirm this is how we want to do hover. (I think this is preferable to, say, using `onMouseEnter` and `onMouseLeave` events and doing this with state, when we can just do with CSS.

## Test Plan

I tried to write a unit test but it's pretty tricky to do this when the actual element isn't built yet. I'll build the test when the actual UI is there.