[BE-1746] new event stubs (#13292)
## Describe your changes

Title! Aligned with Pedro on what we want to show, these are the ones
that are missing. We probably want to discuss - there's a bunch in here
already that we may not use and could possibly remove, but we don't
support these currently and will need to. There's a corresponding FE
change at [PR-329](https://github.com/onefootprint/monorepo/pull/13291)

## Caveats


## Test Plan

CI