Add VDR testing terraform config (#13312)
## Describe your changes
Uploading the Terraform I've been using to test VDR. Not exactly sure where to put this, so I created a new `infra/` directory for smaller standalone config-as-code modules that aren't really part of the `backend/`. One could imagine we add other config as code to this directory, like corp SSO config, etc. I used terraform since it's more of an industry standard so we can provide configs in our VDR docs that we've actually tested. (e.g. Grid just copied and pasted the config in the docs, and didn't have to write their own config.)

## Caveats

## Test Plan