Pass all onboarding session metadata into user auth session (#13344)
## Describe your changes
The "metadata" that we store inside an onboarding session token is growing! Instead of manually adding a new field for each of these pieces of metadata to the user auth session as well, we now just embed the whole metadata inside the user auth session for convenience

TODO DNL: this is a breaking change - it will break any session currently issued with `allow_reonboard = true`. I initially thought this wasn't used, but it actually is for _user-specific_ sessions since this defaults to true for a user-specific session....

## Caveats

## Test Plan