Make vres_id on `RiskSignal` nullable (#13301)
## Describe your changes
Up to this point, `verification_result_id` is non-null. But in order to support step ups in KYB, we need to be able to write the `possible_bo_missing` risk signal _before_ we move to vendor calls and get `vres_ids`. So, here we allow this.

It actually makes life a bit easier for a few places we do this anyways, and is so much nicer now that we have `wf` scoped RSG

## Caveats
will run in shell ofc

## Test Plan
existing/compiler