[BE-1765] Hide header name for FirmEmployeeAssumeAuthContext (#13413)
## Describe your changes
Before this PR, if auth wasn't provided for an API-key-authed endpoint,
we would show an error message that suggested you could provide
`X-Fp-Dashboard-Authorization`. this is because all of our
API-key-authed endpoints _also_ except a firm employee's dashboard auth
token. But, we don't want to expose this detail to customers.

## Caveats


## Test Plan

Before this PR:
```
{"message":"Missing header: X-Footprint-Secret-Key or X-Fp-Dashboard-Authorization","code":"E123","context":{"header":"X-Footprint-Secret-Key or X-Fp-Dashboard-Authorization"},"support_id":"33fb4860-e8b5-4a51-a55c-2ea93230cee6","debug":"MissingHeader("X-Footprint-Secret-Key or X-Fp-Dashboard-Authorization")"}
```
After this PR:
```
$ curl localhost:8000/businesses -X POST
{"message":"Missing header: X-Footprint-Secret-Key","code":"E123","context":{"header":"X-Footprint-Secret-Key"},"support_id":"991b8c1e-8e34-4256-8a3b-915fc33428eb","debug":"MissingHeader(\"X-Footprint-Secret-Key\")"}
```