fix default rules for kyb to not be `not` rules (#13416)
## Describe your changes
`not` rules are really difficult to manage in sandbox.

Now that we are including situations that require executing rules to
simulate stepping up to business documents, we need to have
deterministic behavior of the rule execution. previously, step up tests
were "correctly" passing, but were testing the wrong behavior. this
fixes things

## Caveats


## Test Plan