[BE-1761] Handle errors during initialization of business onboarding (#13404)
## Describe your changes
Now that this happens in a separate hook for `POST /hosted/business/onboarding`, we need the same error handling that we had with `POST /hosted/onboarding` before. Before this PR, we just get a perpetual loading screen upon error.

This starts to use the parent "onboarding requirement" state machine's error handling in order to display when an error occurred while initializing.

There's some special logic to display a custom message for a very specific error that may occur during integration for tenants sending us a `business_external_id`

## Caveats

## Test Plan
Tested locally

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/xncBfoFfDL8XmXxkoXEI/88eca0e9-18a3-4eb8-b296-a650113a8fc1.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/xncBfoFfDL8XmXxkoXEI/88eca0e9-18a3-4eb8-b296-a650113a8fc1.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/88eca0e9-18a3-4eb8-b296-a650113a8fc1.mov">Screen Recording 2024-10-29 at 17.34.32.mov</video>