Don't send vault data for immutable BOs (#13470)
## Describe your changes
We used to have this logic directly in the `useBusinessOwnersPatch` hook but removed it because callers of the hook would never try to update immutable BOs' data. In the next PR, we'll start intermingling immutable BO data a bit more with mutable BO data, so this just re-adds that logic to ensure we don't update immutable BOs' data.

## Caveats
No strong opinion about whether this should be in the hook or in the caller, lmk if you prefer to do it directly in the hook :)

## Test Plan