[BE-1769] Support editing BOs on the confirm screen (#13468)
## Describe your changes
We used to have this functionality but dropped it in #13257 to decrease the scope of the initial PR that migrated to use our new BO APIs in the regular "new business" flow. This finally re-adds that functionality

With a few careful refactors to the BO code in the PRs below this, it makes it _really_ easy to build the confirm screen reusing almost all of our existing code.

## Caveats

## Test Plan
Confirm screen for all mutable BOs:
https://github.com/user-attachments/assets/7ff72ada-0cd6-45f9-8c2d-da8e19201360

Confirm screen when there are immutable BOs (in the "request additional BOs" form):

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/xncBfoFfDL8XmXxkoXEI/9fdee4b1-2419-4d74-a412-66cc7786e961.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/xncBfoFfDL8XmXxkoXEI/9fdee4b1-2419-4d74-a412-66cc7786e961.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/9fdee4b1-2419-4d74-a412-66cc7786e961.mov">Screen Recording 2024-10-30 at 20.55.10.mov</video>