[FE-744] Add util for doc viewer's table of contents to find the currently visible upload (#13394)
## Describe your changes
PR 2/3 for table of contents feature. Adds nothing to UI, is just a
function!

For the table of contents, the "active" upload (or the one that's
current being viewed), is highlighted in purple:
![Screenshot 2024-10-29 at
21.31.04.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/9c0BPEPR5TrhOpgYx9RG/d0dcc5ea-b473-475b-9971-90220ef4556a.png)

This function checks the element bounds in relation to the parent's to
see which upload is currently in frame. It prioritizes the upload that's
entirely in frame, then checks if there's an upload thats center is in
frame.

## Caveats


## Test Plan