[FP-746] Add util to doc viewer to determine whether to show a mini preview of the selfie (#13419)
## Describe your changes
PR [1/2] for the ext nice-to-have doc viewer v2 feature: a
miniature/minimized selfie preview in the bottom right corner of the
LicenseDetails modal. Title! This adds nothing to the UI, is just a util
that determines when to show it. We only want to show the selfie if it's
a successful upload and if the selfie isn't currently in frame (ie we're
looking at the driver's license, then selfie preview should show in the
bottom corner, but if we're looking at the selfie the preview shouldn't
show)

How it looks in the figma:

![Screenshot 2024-10-30 at
12.23.05.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/9c0BPEPR5TrhOpgYx9RG/e9544d27-298a-4612-b860-54f03f2f6306.png)

## Caveats


## Test Plan