[PR-333] timeline component scaffolding (#13336)
## Describe your changes

Realized that to do this `Line` change, my initial structure wasn't going to work so well - this basically makes the changes so we can use something almost directly taken from the previous `Timeline` component, with some changes for data handling. I'll change some of the styling here in the future, this is just general scaffolding.

![Screenshot 2024-10-28 at 3.10.18 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/53ae1db3-1bb9-4755-a709-93b791182e4f.png)

## Caveats

## Test Plan