[PR-349] AccessEventDetail type (#13464)
## Describe your changes

Mirrors the structure on the backend — it's quite tricky to type these
`AccessEventDetail`s because they vary pretty wildly in structure. I see
this more as a proposed fix - let me know what you think / if this looks
broadly reasonable! Can easily replicate this structure as we add more
and more of the BE detail types in.

## Caveats


## Test Plan

Typecheck / existing tests