[BE-1776] Bugfix: stop requiring business selection for document workflows (#13545)
## Describe your changes
There was a bug here for a specific case: if a document was requested
from a user who previously onboarded onto a KYB playbook, the document
workflow would be associated with the KYB playbook. We would erroneously
assume that a business needed to be created / selected in order to
upload the document for the playbook. This fixes that

Arg playbooks associated with document workflows 😠 

## Caveats


## Test Plan
Integration test. It failed before this code change, passes after