[PR-350] new principal type (#13465)
## Describe your changes


Title. Mirrors the changes to the `Actor` serializer on the BE that we
push in
[BE-1740](https://linear.app/footprint/issue/BE-1740/create-org-role-add-name-metadata)
and allows for the correct display of name and email that we add in
[PR-344]().

This now also adds UI so that we actually display these properly - was
going to batch this into two, but the changes are so interrelated /
would've had to build a weird intermediate step that I did them in one.
The bulk of this is tests (and the test fixtures) in
`principal-actor.test.config.ts`


Might still be more casting than we'd like but I think this is the best
solution for now (or: I'm open to others! Curious what you both think)

![Screenshot 2024-10-30 at
7.05.07 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/28d21647-b732-482c-86f1-8cf4491ea7ff.png)

## Caveats


## Test Plan
Typecheck