[PR-353] align dot and event title (#13548)
## Describe your changes

Title! This is a weird PR because I'd love if we could figure out a better solution, but Pedro says it looks good so I think we should keep. Problem is there's no good way to align the top of the event text with the dot (and date) in the case that the event spans more than two lines. If it is two lines, and we try to center align, the dot and date will be aligned in the middle of the two lines, which is no good. This fixes it manually with some padding and looks good for those cases. If you have other ideas, all ears!

<img width="940" alt="Screenshot 2024-10-31 at 6 17 05 PM" src="https://github.com/user-attachments/assets/0a70dc7a-6477-4baf-ad6e-e96f11a54631">

## Caveats

## Test Plan