[1/3] Expose all user bootstrap data from BO token (#13538)
## Describe your changes
The frontend will start bootstrapping all of this in.

## Caveats
My longer-term vision is that the bootstrap data doesn't need to be
passed down client side and can instead be stored in a server-side token
that doesn't leak it to the client. Some functionality _kind_ of like
this exists in the ob session tokens. But I could also see this _maybe_
in the user tokens?? Maybe not.

For now, this just additionally exposes the name attributes so they'll
be prefilled for the secondary BO

## Test Plan
integration test