Create indexes to speed up audit event filter query (#13636)
## Describe your changes
The slow part of `AuditEvent::filter` is the filter on
`audit_event::is_live.eq(params.is_live).or(audit_event:;is_live.is_null())`.

The trick here is that `is_live` is an optional bool, so we can
enumerate all the possible filters (just two -- `is_live = true or
is_live is null` and `is_live = false or is_live is null`). Then we can
create an index for each of those expressions. We can't use this trick
on e.g. a nullable string, like the similar OR problem we have in VDR.
But it works here! I tested these new indexes in prod already and they
make the page loads noticeably faster.

## Caveats


## Test Plan
Applied in dev and prod already