[PR-373] role display (#13629)
## Describe your changes

This is much like Principal actor - note that we won't have this weird overlap because we'll never show a role (actually) in a decryption event. This is just to test th e UI looks sensible. Also realized I could nest some previous stuff here as a component, so lots of file moves in this PR.

![Screenshot 2024-11-01 at 2.24.55 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/6d85502c-4ece-44aa-9781-61bed10a263f.png)

## Caveats

## Test Plan

In UI and `role-display.test.tx`